+++ lib/src/models/goods_and_services_tax.dart
@@ -1,0 +1,18 @@
+class GoodsAndServicesTax {
+  final String rate;
+  final String type;
+
+  GoodsAndServicesTax({required this.rate, required this.type});
+
+  factory GoodsAndServicesTax.fromCBJson(Map<String, dynamic> json) {
+    return GoodsAndServicesTax(
+      rate: json['rate'],
+      type: json['type'],
+    );
+  }
+
+  @override
+  String toString() {
+    return 'GoodsAndServicesTax{rate: $rate, type: $type}';
+  }
+}