+++ test/rest/fees/fees_test.dart
@@ -1,0 +1,51 @@
+import 'dart:io';
+
+import 'package:coinbase_cloud_advanced_trade_client/src/models/credential.dart';
+import 'package:coinbase_cloud_advanced_trade_client/src/rest/fees.dart';
+import 'package:http/http.dart' as http;
+import 'package:mockito/annotations.dart';
+import 'package:mockito/mockito.dart';
+import 'package:test/test.dart';
+
+import '../../test_constants.dart';
+import '../../tools.dart';
+import './fees_test.mocks.dart';
+
+final Map<String, String> envVars = Platform.environment;
+final String? skipTests = envVars['SKIP_TESTS'];
+final bool skip = skipTests == 'false' ? false : true;
+
+final Credential credentials =
+    Credential(apiKeyName: apiKeyName, privateKeyPEM: privateKeyPEM);
+
+@GenerateMocks([http.Client])
+void main() {
+  group('Fees REST Tests using Mocks', () {
+    test('Get Transaction Summary', () async {
+      final client = MockClient();
+      final String getTransactionSummaryJson =
+          await getJsonFromFile('rest/fees/get_transaction_summary.json');
+
+      when(client.get(any, headers: anyNamed('headers'))).thenAnswer(
+          (_) async => http.Response(getTransactionSummaryJson, 200));
+
+      final result =
+          await getTransactionSummary(client: client, credential: credentials);
+
+      expect(result, isNotNull);
+      expect(result!.totalFees, isNotNull);
+    });
+  });
+
+  group('Fees REST Tests Requests to Coinbase AT API Endpoints', skip: skip,
+      () {
+    test('Get Transaction Summary', () async {
+      final result = await getTransactionSummary(credential: credentials);
+
+      print('Fees: $result');
+
+      expect(result, isNotNull);
+      expect(result!.totalFees, isNotNull);
+    });
+  });
+}